pyproject.toml
[project]
name = "microservice-order"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiokafka>=0.12.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "pre-commit>=4.5.0",
    "psycopg[binary]>=3.2.13",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.15.4",
    "ruff>=0.14.7",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
]
